combine jvm/android socket listeners, catch exceptions
